backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
python-jose[cryptography]==3.5.0
bcrypt==4.3.0
celery[redis]==5.5.3
redis==5.2.1
pydantic-settings==2.11.0
email-validator==2.3.0
python-multipart==0.0.20
httpx==0.28.1
pdfplumber==0.11.4
python-docx==1.1.2
